feat(F-207): restructure components folder
