+++ frontend/src/app/[locale]/(about)/components/AboutPage.tsx
