+++ frontend/src/app/[locale]/(main)/admin/page.tsx
@@ -1,7 +1,7 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/MetadataProps';
-import Admin from '@/components/common/AdminPage';
+import Admin from './components/AdminPage';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
   const { locale } = await params;
