+++ frontend/src/app/[locale]/(main)/dashboard/components/HistoryItem.tsx
