+++ frontend/src/app/[locale]/(main)/feedback/[id]/components/FeedbackDetail.tsx
