+++ frontend/src/app/[locale]/(main)/feedback/[id]/components/FeedbackQuote.tsx
