+++ frontend/src/app/[locale]/(main)/feedback/[id]/page.tsx
@@ -1,7 +1,7 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
-import FeedbackDetail from '@/components/common/FeedbackDetail';
 import { MetadataProps } from '@/interfaces/MetadataProps';
+import FeedbackDetail from './components/FeedbackDetail';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
   const { locale } = await params;
