+++ frontend/src/app/[locale]/(main)/history/components/HistoryStats.tsx
