+++ frontend/src/app/[locale]/(main)/history/components/PreviousSessions.tsx
