+++ frontend/src/app/[locale]/(main)/history/page.tsx
@@ -2,8 +2,8 @@ import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/MetadataProps';
 import HistoryHeader from '@/components/layout/HistoryHeader';
-import HistoryStats from '@/components/common/HistoryStats';
-import PreviousSessions from '@/components/common/PreviousSessions';
+import HistoryStats from './components/HistoryStats';
+import PreviousSessions from './components/PreviousSessions';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
   const { locale } = await params;
