+++ frontend/src/app/[locale]/(main)/new-training/components/NewTrainingCustomizeStep.tsx
