+++ frontend/src/app/[locale]/(main)/new-training/components/NewTrainingForm.tsx
