+++ frontend/src/app/[locale]/(main)/new-training/components/NewTrainingSituationStep.tsx
