+++ frontend/src/app/[locale]/(main)/preparation/components/ObjectivesList.tsx
