+++ frontend/src/app/[locale]/(standalone)/simulation/[id]/page.tsx
@@ -1,7 +1,7 @@
-import SimulationPageComponent from '@/components/common/SimulationPage';
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/MetadataProps';
+import SimulationPageComponent from './components/SimulationPage';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
   const { locale } = await params;
